agadfgafg

+++ web/app/plugins/sttv/includes/functions/sttv-webhook-functions.php
@@ -20,8 +20,8 @@ function trial_expiration_checker() {
 
     if ( !empty( $garbage ) ) {
         foreach ( $garbage as $g ) {
-            if ( $g['exp_date'] > 0 ) {
-                $umeta = get_user_meta( $g['wp_id'], 'sttv_user_data', true );
+            $umeta = get_user_meta( $g['wp_id'], 'sttv_user_data', true );
+            if ( $g['exp_date'] > 0 && $umeta ) {
                 try {
                     $customer = \Stripe\Customer::retrieve( $umeta['customer'] );
                     $customer->delete();
